refactor(editor): split icons into individual files, regenerate lockfile
One TSX component per icon file under src/ui/icons/. No shared
abstractions — each file is a self-contained SVG component.
