+++ packages/editor/src/ui/icons/index.ts
@@ -1,0 +1,29 @@
+export { AlignCenterIcon } from './align-center';
+export { AlignLeftIcon } from './align-left';
+export { AlignRightIcon } from './align-right';
+export { BoldIcon } from './bold';
+export { CaseUpperIcon } from './case-upper';
+export { Check } from './check';
+export { ChevronDown } from './chevron-down';
+export { Code, CodeIcon } from './code';
+export { Columns2 } from './columns-2';
+export { Columns3 } from './columns-3';
+export { Columns4 } from './columns-4';
+export { ExternalLinkIcon } from './external-link';
+export { Heading1 } from './heading-1';
+export { Heading2 } from './heading-2';
+export { Heading3 } from './heading-3';
+export { ItalicIcon } from './italic';
+export { LinkIcon } from './link';
+export { List } from './list';
+export { ListOrdered } from './list-ordered';
+export { MousePointer } from './mouse-pointer';
+export { PencilIcon } from './pencil';
+export { Rows2 } from './rows-2';
+export { SplitSquareVertical } from './split-square-vertical';
+export { SquareCode } from './square-code';
+export { StrikethroughIcon } from './strikethrough';
+export { Text, TextIcon } from './text';
+export { TextQuote } from './text-quote';
+export { UnderlineIcon } from './underline';
+export { UnlinkIcon } from './unlink';
