+++ packages/editor/src/ui/icons/mouse-pointer.tsx
@@ -1,0 +1,26 @@
+import type * as React from 'react';
+
+interface IconProps extends React.SVGAttributes<SVGSVGElement> {
+  size?: number | string;
+}
+
+export function MousePointer({ size, width, height, ...props }: IconProps) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width={size ?? width ?? 24}
+      height={size ?? height ?? 24}
+      viewBox="0 0 24 24"
+      fill="none"
+      stroke="currentColor"
+      strokeWidth={2}
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      aria-hidden="true"
+      {...props}
+    >
+      <path d="M12.586 12.586 19 19" />
+      <path d="M3.688 3.037a.497.497 0 0 0-.651.651l6.5 15.999a.501.501 0 0 0 .947-.062l1.569-6.083a2 2 0 0 1 1.448-1.479l6.124-1.579a.5.5 0 0 0 .063-.947z" />
+    </svg>
+  );
+}
