+++ packages/editor/src/ui/icons/pencil.tsx
@@ -1,0 +1,26 @@
+import type * as React from 'react';
+
+interface IconProps extends React.SVGAttributes<SVGSVGElement> {
+  size?: number | string;
+}
+
+export function PencilIcon({ size, width, height, ...props }: IconProps) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width={size ?? width ?? 24}
+      height={size ?? height ?? 24}
+      viewBox="0 0 24 24"
+      fill="none"
+      stroke="currentColor"
+      strokeWidth={2}
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      aria-hidden="true"
+      {...props}
+    >
+      <path d="M21.174 6.812a1 1 0 0 0-3.986-3.987L3.842 16.174a2 2 0 0 0-.5.83l-1.321 4.352a.5.5 0 0 0 .623.622l4.353-1.32a2 2 0 0 0 .83-.497z" />
+      <path d="m15 5 4 4" />
+    </svg>
+  );
+}
