+++ packages/editor/src/ui/icons/split-square-vertical.tsx
@@ -1,0 +1,32 @@
+import type * as React from 'react';
+
+interface IconProps extends React.SVGAttributes<SVGSVGElement> {
+  size?: number | string;
+}
+
+export function SplitSquareVertical({
+  size,
+  width,
+  height,
+  ...props
+}: IconProps) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width={size ?? width ?? 24}
+      height={size ?? height ?? 24}
+      viewBox="0 0 24 24"
+      fill="none"
+      stroke="currentColor"
+      strokeWidth={2}
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      aria-hidden="true"
+      {...props}
+    >
+      <path d="M5 8V5c0-1 1-2 2-2h10c1 0 2 1 2 2v3" />
+      <path d="M19 16v3c0 1-1 2-2 2H7c-1 0-2-1-2-2v-3" />
+      <line x1="4" x2="20" y1="12" y2="12" />
+    </svg>
+  );
+}
