+++ packages/editor/src/ui/icons/square-code.tsx
@@ -1,0 +1,27 @@
+import type * as React from 'react';
+
+interface IconProps extends React.SVGAttributes<SVGSVGElement> {
+  size?: number | string;
+}
+
+export function SquareCode({ size, width, height, ...props }: IconProps) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width={size ?? width ?? 24}
+      height={size ?? height ?? 24}
+      viewBox="0 0 24 24"
+      fill="none"
+      stroke="currentColor"
+      strokeWidth={2}
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      aria-hidden="true"
+      {...props}
+    >
+      <path d="m10 9-3 3 3 3" />
+      <path d="m14 15 3-3-3-3" />
+      <rect x="3" y="3" width="18" height="18" rx="2" />
+    </svg>
+  );
+}
