+++ packages/editor/src/ui/icons/strikethrough.tsx
@@ -1,0 +1,32 @@
+import type * as React from 'react';
+
+interface IconProps extends React.SVGAttributes<SVGSVGElement> {
+  size?: number | string;
+}
+
+export function StrikethroughIcon({
+  size,
+  width,
+  height,
+  ...props
+}: IconProps) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width={size ?? width ?? 24}
+      height={size ?? height ?? 24}
+      viewBox="0 0 24 24"
+      fill="none"
+      stroke="currentColor"
+      strokeWidth={2}
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      aria-hidden="true"
+      {...props}
+    >
+      <path d="M16 4H9a3 3 0 0 0-2.83 4" />
+      <path d="M14 12a4 4 0 0 1 0 8H6" />
+      <line x1="4" x2="20" y1="12" y2="12" />
+    </svg>
+  );
+}
